GH-1203: Extract ContainerProperties Super Class
Resolves https://github.com/spring-projects/spring-kafka/issues/1203

Common properties that can be used with any consumer, not just a consumer that's
in a container.
